refactor(quick-260504-my5): split ConfidenceTooltipContent into WDL and eval variants; correct test name in eval tooltip
Tooltip body component had two unrelated rendering paths gated by a single
prop. Split into WdlConfidenceTooltip (binomial-style score test) and
EvalConfidenceTooltip (eval mean vs baseline) so each call site only sees
the props it needs.

Eval tooltip layout:
- raw eval and 95% CI on one line
- centered eval and 95% CI on the next line (delta vs the per-color
  engine baseline that the bullet chart centers on)

Also corrected the eval-tooltip footnote and probability copy: backend
uses a two-sided Wald z-test against the per-color baseline (+0.315 pawns
white, -0.189 black) -- not a one-sample t-test vs 0.

BulletConfidencePopover now takes a required centerPawns prop; wired
through MostPlayedOpeningsTable (desktop) and Openings.tsx (mobile).
ConfidencePill is now WDL-only -- no production callers used eval mode.
