+++ frontend/src/pages/Openings.tsx
@@ -218,6 +218,7 @@ function MobileMostPlayedRows({
                           evalMeanPawns={o.avg_eval_pawns}
                           evalCiLowPawns={o.eval_ci_low_pawns}
                           evalCiHighPawns={o.eval_ci_high_pawns}
+                          centerPawns={evalBaselinePawns}
                           testId={`${testIdPrefix}-bullet-popover-mobile-${rowKey}`}
                           prefaceText={buildMgEvalHeaderTooltip(evalBaselinePawns)}
                         />
